% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascertainment.R
\name{ascertainment}
\alias{ascertainment}
\title{Estimate case ascertainment ratio}
\usage{
ascertainment(cfr_distr, death_to_case)
}
\arguments{
\item{cfr_distr}{}

\item{death_to_case}{}
}
\value{

}
\description{
Estimate case ascertainment ratio
}
