% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{augment_cases}
\alias{augment_cases}
\title{Augment the time series of cases}
\usage{
augment_cases(cases, previous, ndays, nsim = 10000)
}
\value{
Matrix of observed and augmented cases. Dimensions of this
matrix are (2 * nsim) X (previous) where the observed values are
repeated across
}
\description{
Augment the time series of cases
for the days ahead to be able to
project deaths.
}
\details{
Augmentation of case data for the
next \code{ndays} days involves
drawing samples from a distribution
(here \code{gamma} distribution)
with parameters estimated from the
last \code{previous} days.
}
\author{
Sangeeta Bhatia
}
