% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_epidemic_size.R
\name{episize_after_mu}
\alias{episize_after_mu}
\title{Estimate epidemic size from the present time  minus the mean
delay between case and death up to present}
\usage{
episize_after_mu(cases, rho)
}
\value{

}
\description{
Estimate epidemic size from the present time  minus the mean
delay between case and death up to present
}
\author{
Sangeeta Bhatia
}
