% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_epidemic_size.R
\name{episize_before_mu}
\alias{episize_before_mu}
\title{Estimate size of the epidemic upto the present time minus the mean
delay from case to death.}
\usage{
episize_before_mu(deaths, mu_delta, cfr_distr)
}
\value{

}
\description{
Estimate size of the epidemic upto the present time minus the mean
delay from case to death.
}
\author{
Sangeeta Bhatia
}
