% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_deaths}
\alias{project_deaths}
\title{Project the number of deaths}
\usage{
project_deaths(augmented_cases, weights, trunc, ndays, nsim, deaths_to_cases)
}
\value{

}
\description{
Project the number of deaths
}
\author{
Sangeeta Bhatia
}
