% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_deaths_cases.R
\name{ratio_deaths_cases}
\alias{ratio_deaths_cases}
\title{Estimate ratio if deaths to reported cases}
\usage{
ratio_deaths_cases(wtd_incid, deaths, nsamples = 10000, twindow = 7)
}
\arguments{
\item{wtd_incid}{}

\item{deaths}{}

\item{nsamples}{}

\item{twindow}{}
}
\value{

}
\description{
Estimate ratio if deaths to reported cases
}
