% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_incid.R
\name{weighted_incid}
\alias{weighted_incid}
\title{Return the incidence time-series weighted by the reporting to
death delay}
\usage{
weighted_incid(incid, weights, trunc)
}
\arguments{
\item{incid}{incidence time series as a T X 1 matrix}

\item{weights}{Discrete probability distribution of the reporting
to death delay}
}
\value{

}
\description{
Returns
}
\author{
Sangeeta Bhatia
}
